constructor - deprecated standard
This is because of a deprecated standard. In the file, replace the line:

function Election () public {
with:
constructor() public {
The line of code shown just above is a constructor, it runs on a contract's deployment, and it is used(as in the file in context) to save the contract owner's address(msg.sender varies based on who interacts with the contract).
This replaces the existing implementation of using a function with the same name as the contract to act as a constructor with a separate constructor function.